src/AzureFunc/requirements.txt
azure-functions==1.0.7
joblib==0.14.1
